+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1 (unreleased)
 
 * Library extracted from Monolith
+* AWS Lambda and Aleph Monolith support